+++ packages/uni-app/uni_modules/vite-plugin/js_sdk/common/index.d.ts
@@ -1,6 +1,7 @@
 export { DateFormatOptions, formatDate, formatFileSize, getDateFormatParams, parseTemplate } from './format/index.js';
 export { CopyOptions, CopyResult, ScanDirectoryOptions, ScannedFile, checkSourceExists, copySourceToTarget, scanDirectory, shouldUpdateFileContent, writeFileContent, writeFileSyncSafely, writeJsonReport } from './fs/index.js';
 export { SanitizeRuleOptions, escapeHtmlAttr, injectBeforeTag, injectHeadAndBody, sanitizeContent } from './html/index.js';
+export { isExtensionIncluded, isPathExcluded, isPreCompressed, normalizePath } from './path/index.js';
 export { makeCallback } from './script/index.js';
 export { ANSI } from './ui/index.js';
 export { V as Validator } from '../shared/vite-plugin.DRRlWY8P.js';